requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.6.0
asyncio==3.4.3
attrs==23.2.0
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
frozenlist==1.4.1
idna==3.4
multidict==6.0.5
python-decouple >= 3.8
pymongo >= 3.11.4
pytz==2022.7.1
requests==2.28.2
soupsieve==2.4
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.15
yarl==1.9.4